moves AddTodo into own component file - App imports AddTodo
